카카오 맵 컴포넌트를 새로 만들어 관련 함수 및 태그 이동
